% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcharts_geo_process.R
\name{gchart_available_areas}
\alias{gchart_available_areas}
\title{Get available areas}
\usage{
gchart_available_areas(
  type = c("countries", "us-areas"),
  cache = gchart_get_cache_path()
)
}
\arguments{
\item{type}{Type of geo-resource. Defaults to countries.}

\item{cache}{Path where the downloaded data is stored.}
}
\description{
Get a list of areas that have been downloaded
}
