% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gchart_cache_dir}
\alias{gchart_cache_dir}
\title{Verify cache directory}
\usage{
gchart_cache_dir(
  path = tools::R_user_dir(package = "gchartsmap", which = "cache")
)
}
\arguments{
\item{path}{Path to verify}
}
\description{
Ensure that the cache directory exists
}
