% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gchart_get_cache_path}
\alias{gchart_get_cache_path}
\title{Get the cache path}
\usage{
gchart_get_cache_path(path = NULL)
}
\arguments{
\item{path}{Path to use as cache}
}
\value{
Returns the path to the local cache as set in the
`R_GOOGLE_CHART_CACHE` environment variable. If that is not set, gets the
system's default cache path for the package as provided by
`tools::R_user_dir()`.
}
\description{
Get the saved cache path
}
\examples{

library(package = "gchartsmap")

# set the cache path to a temp folder
gchartsmap::gchart_set_cache(path = tempdir())

# check the set cache
gchartsmap::gchart_get_cache_path()

}
