% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcharts_countries.R
\name{gchart_countries}
\alias{gchart_countries}
\title{Extracts the list of available countries from the Google Charts documentation}
\usage{
gchart_countries(
  url = "https://developers.google.com/chart/interactive/docs/gallery/geochart",
  cache = FALSE,
  path = NULL,
  format = NULL
)
}
\arguments{
\item{url}{Location for the relevant Google Charts documentation.
Use the default.}

\item{cache}{Whether the countries table should be stored in disk}

\item{path}{Place to store the countries table if it is being cached}

\item{format}{Which format to use for storing the countries table}
}
\value{
A \code{data.frame} with information for all countries available in
the service.
}
\description{
Extracts the list of available countries from the Google Charts documentation
}
\details{
Extracts the information of countries available in the Google Charts geochart
service from the online documentation
}
