% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcharts_us_areas.R
\name{gchart_generate_us_areas}
\alias{gchart_generate_us_areas}
\title{Generate Google Charts spatial data for US areas}
\usage{
gchart_generate_us_areas()
}
\description{
This function queries Google Charts resources to identify the US geographic
areas used in services like Google Trends, and uses geographic data from the
US Census Bureau to provide those areas with subdivisions at the county level.
}
