% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gchart_get_cache_path}
\alias{gchart_get_cache_path}
\title{Get the cache path}
\usage{
gchart_get_cache_path(path = NULL)
}
\arguments{
\item{path}{Path to use as cache}
}
\description{
Get the saved cache path
}
