% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcharts_global_areas.R
\name{gchart_get_countries}
\alias{gchart_get_countries}
\title{Get Google Charts data for countries}
\usage{
gchart_get_countries(
  countries = "all",
  server = "https://www.gstatic.com/charts/geochart/10/mapfiles/",
  cache = gchart_get_cache_path()
)
}
\arguments{
\item{countries}{Country codes to get.}

\item{server}{Google geochart server to access.}

\item{cache}{Path to store downloaded data.}
}
\description{
Access the Google Charts geochart data for countries
}
\details{
The function invisibly returns the file path for successful requests or
the response status code for failed requests, in a character vector with the
country code for each element.
}
