% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcharts_us_areas.R
\name{gchart_process_areas}
\alias{gchart_process_areas}
\title{Process Google Charts data for US areas}
\usage{
gchart_process_areas(
  areas,
  type = c("countries", "us-areas"),
  cache = gchart_get_cache_path()
)
}
\arguments{
\item{areas}{Area codes to get. Should be integers. If not provided, all
available areas are processed.}

\item{type}{Type of geo-resource. Defaults to countries.}

\item{cache}{Path where the downloaded data is stored.}
}
\description{
Process the downloaded Google Charts geochart data for US areas
}
\details{
Google Charts data is served as JavaScript code that defines objects with the
desired data. This function processes the locally-available Google Charts js
files to generate spatial objects.
}
