% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gchart_remove_cache_path}
\alias{gchart_remove_cache_path}
\title{Delete a saved cache path}
\usage{
gchart_remove_cache_path(remove = FALSE, path = NULL)
}
\arguments{
\item{remove}{Whether to remove the path.}

\item{path}{Path to look for an .Renviron file.}
}
\description{
This function will uninstall the cache path from the environment
variables. If a path is provided, it will
be used to remove the variable from .Renviron file in that path.
}
\keyword{internal}
