% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gchartsmap-package.R
\docType{package}
\name{gchartsmap-package}
\alias{gchartsmap}
\alias{gchartsmap-package}
\title{gchartsmap: Access 'Google Charts' Map Data}
\description{
Connects to the 'Google Charts' geographic data resources described in \url{https://developers.google.com/chart/interactive/docs/gallery/geochart}, allowing the user to download contents to use as a reference for related services like 'Google Trends'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/odeleongt/gchartsmap}
  \item Report bugs at \url{https://github.com/odeleongt/gchartsmap/issues}
}

}
\author{
\strong{Maintainer}: Oscar de Leon \email{odeleon@emory.edu} (\href{https://orcid.org/0000-0003-1344-4412}{ORCID}) [copyright holder]

}
\keyword{internal}
