% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{path_existing_root}
\alias{path_existing_root}
\title{Find existing path}
\usage{
path_existing_root(path)
}
\arguments{
\item{path}{Path to analyze.}
}
\description{
Find the deepest path component that exists.
}
\keyword{internal}
